Fix presence timeout issue.
The server was not receiving the "Are you alive?" ping from the client
on presence check because token length was not included by the client.
Added token length to fix that.

The client app was timing out on presence notfications even though
the server was online. We were not resetting the client timer when we
get the servers "I am alive" message. Added resetting timeout logic.

Change-Id: Ibded89835e62cc283882c897ffe98c0fbde7f12c
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/759
Reviewed-by: Erich Keane <erich.keane@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>