Add an exception handling logic in collection presence in Group Manager
subscribeCollectionPresence() in GroupManager is intended to be used for group
resource having remote resources as child resources, not conventional
collection resource having local resources.
After requesting a list of child resources by sending GET to a collection
resource, if the child resource has a URI without a coap prefix "coap://", \
it is obviously a local (child) resource, which is not a desired resource
we want to send a request to presence check. Thus, we're going to ignore it.

Ref: [IOT-289]

Change-Id: I517f4fc25bb6a5e964e3624eb858a15c2dbafc03
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/773
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>