Added new function in EntityHandler to parse and validate query string.
IOT-491: When client creates a resource and calls GET  with different
resource type name and interface,GET should fail but it is successful.
Analysis: Since APP creates resource, it should know the 
resource/interface types and it should check for resourceTypeName
and ResourceInterface to identify a known resource prior to handle
entity Request.

Change-Id: I9159b465bf708a4cb5835f4b3adf0aea3a2c87bb
Signed-off-by: Moumita Ray <moumita.ray@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/860
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>