Update C++ samples to work after WiFi/Ethernet adapter merge.
The merge of ethernet and wifi includes a check in the adapter for whether
a request is going out to a multicast or unicast endpoint.
The decision for whether it is a multicast or unicast request is made on
the basis of whether the group endpoint is set or the remote endpoint.

The C stack creates a group endpoint if the connectivity type is
OC_ALL or a remote endpoint otherwise. But all connectivity types are mapped
to CA_IPV4. To ensure that a group endpoint is created for multicastm the C++
samples should make the findResource() call with OC_ALL.
The new constant is the multicast resource discovery uri

Change-Id: I8933a92a18ec920cf9b4b7dd994a6cbd82cb1c44
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/898
Reviewed-by: Sakthivel Samidurai <sakthivel.samidurai@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>