Force linking of liboctbstack to require all symbols resolved
The --no-undefined option (same as -z defs) causes the linker to refuse
to create its output if there are any symbols still unresolved after
searching all the dependent shared libraries. The default on ELF systems
is to allow shared libraries to link even if there are symbols
undefined, which has caused many a developer a lot of headache.

Unfortunately, changing ELF design flaws is not within our grasp.

Change-Id: I0d4913955e3745b69672ffff13dbc5f36ac07b3d
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/921
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>