Changed ServerInstanceID to UUID
The specification and design document both specify that server instance
ID should be a UUID, yet this was initially implemented as a Uint32.

This patch implements UUID generation, UUID-as-string generation,
and UUID-to-string conversion in the ocrandom module.  Additionally,
the server instance ID calls now correctly work with UUID rather than
the int32.  OCResourceIdentifer also now work directly with strings
rather than converting the value to an int32.

Change-Id: Ic37559973f82515f795d75f798dab6b6ff99d31f
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/983
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>
Reviewed-by: Doug Hudson <douglas.hudson@intel.com>
Reviewed-by: Joseph Morrow <joseph.l.morrow@intel.com>