Fix Android build failures, enable random  unit tests
It was discovered that the ocrandom uuid changes broke the android
build.  This is because Android, despite being an incomplete linux
implementation, defines __linux__.  This patch fixes that, and properly
enables the ocrandom unit tests.

Additionally, the OC_WIFI/OC_ETHERNET removals weren't propogated
properly to the android build, this patch does that in order to fix the
build.

Change-Id: Iabfeb704d046fbadf1fd368259d0e96ca1e74d90
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1020
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-by: Vijay Kesavan <vijay.s.kesavan@intel.com>