Implementation of OICStrcat/OICStrcpy
Repeated mistakes based on copy and cat operations in the C
code shows that we need a unified implementation that is as safe as
possible.  These implementations closely emulate the Microsoft _S
versions of these functions, which seem to be safest feature wise.

Replacing copy and cat operations is reserved for a different
commit.

Change-Id: Id1f6f4515a00679661cc722419fa0bab84a624c2
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1115
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: John Light <john.j.light@intel.com>