CBOR Integration, first functional version
This works properly with all occlient/ocserver tests, with the exception
of Presence, which is broken in master. That fix will be cherry-picked
here when that happens.

There are a large number of TODOs that need to be taken care of, along
with samples needing updating, and the C++ conversions need to happen
properly as well.

Change-Id: Id49df866952fd20eab511ad2d56e4c430d2dab19
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1364