Fix segfault on GET for device and platform info.
The mandatory fields for device and platform are not enforced by the
stack on init. They are enforced only when the info is explicitly added.
Servers that did not register this device and platform info segfaulted
when getting discovery requests from clients.
Added a null check before dereferencing a pointer to fix the problem.

Change-Id: I8b77d149014329e60f8aa8e9c1efb761394487da
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1285
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sakthivel Samidurai <sakthivel.samidurai@intel.com>
Reviewed-by: Erich Keane <erich.keane@intel.com>