Added documentation build script for Doxygen and JavaDoc
This change set adds a script (./build_docs.sh) to generate
the documenation for iotivity which is placed into the ./docs
directory. Currently, the script builds java and C/C++ API
documentation and includes checks for warnings and errors.
When all of the existing warnings and errors are addressed,
this should be added to the build script to fail validation
if/when a contributor breaks the API documentation.

Please consider cherry-picking to the 0.9.1 release branch.

Change-Id: I490fbfc83cf61c76c7f59ba2c00404a073b039e7
Signed-off-by: Patrick Lankswert <patrick.lankswert@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1418
Reviewed-by: Erich Keane <erich.keane@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>