csdk: Fixes issue reported by cppcheck tool
This patch fixes the the below issue reported by cppcheck tool:
"(error) Possible null pointer dereference: ReqInfo - otherwise it is
redundant to check if ReqInfo is null at line 234".

Null value check of ReqInfo is removed as it in unnecessary here, as this check is
already done and OIC_LOG_BUFFER is just printing the content of ReqInfo->info.token.

Change-Id: I8c15d8b582135c445fe89350dc67812e531c52d3
Signed-off-by: Ravi Nanjundappa <nravi.n@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1507
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>