Fixed broken presence after IPv6 plumbing merge to master.
Presence wasn't working for multiple reasons. oc/presence was not moved
to oic/ad, probably lost in the merge to master.
resourceUri in CA data structures was moved form the endpoint to the
CAInfo struct. But wasn't included in a clone operation which resulted
in a race condition when the memory was free'd in the main thread before
being accessed by the thread in CA.
Other adapter/flag issues.
Unicast presence failed to work because host() from OCResource only
returned the IP without the port.

Change-Id: If2e46f4677ebe369cbaee9609de38c369bf4c524
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Signed-off-by: Erich Keane <erich.keane@verizon.net>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1424
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>