resource-manipulation: libserver_builder.so add missing depending library
libserver_builder.so is also depending on liboc.so and
librcs_common.so, this patch adds these dependencies to the build.

liboctbstack.so is only needed when LOGGING=true is set. Without
liboctbstack.so libserver_builder.so misses the dependencies to the
OCLog and OCLogv symbols.
Add LIBPATH to make sure the libs will be found, this fixes a problem
in some jenkins tests.

pthread is added two times, once should be enough.

Without this the dynamic loader will not automatically load these
libraries and it could result in unresolved dependencies at runtime.

Change-Id: I6af126eb4af975b7231d01df922885c7dec5fc56
Signed-off-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1799
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>