Updated auto_build.sh for continuous integration improvements.
This commit effectively adds a new mode of building in the build
script. This new mode can be referenced as "parallel". Before all
builds occurred in sequence and could take a long time. Now an
individual contributor may build just a shim of the entire build
list. This will allow the Jenkins CI system to build each of the
shims in parallel to better utilize the several executors found
across the many build servers we have available for CI.

Change-Id: I945bc45201d8c3d4a5c754de02eaa623f59c8449
Signed-off-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1818
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1964
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>