csdk: libconnectivity_abstraction.so: add missing depending library
libconnectivity_abstractionso depends on libcoap.so and librt.so when
build as shared lib, which is the case for the Tizen build. This patch
adds these dependencies to the build.

Without this the dynamic loader will not automatically load these
libraries and it could result in unresolved dependencies at runtime.

Change-Id: Ib972be2b766d7fbccac9b246ae6f11e6182af94c
Signed-off-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1825
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: Uze Choi <uzchoi@samsung.com>