Corrected files with doxygen comments that were generating warnings.
Fixed several files with out of date or incorrect doxygen comments.
These included use of explicit tags (@fn, @var, @struct, etc.),
incorrect use of "[IN]" and "[OUT]", and other issues.

Aside from the problems actually marked in the output as warnings,
these issues were causing a large portion of the documentation to
be dropped silently.

Change-Id: I66027f4f98a6e82b7db826693b5ec1b1f30e58cf
Signed-off-by: Jon A. Cruz <jonc@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1836
Reviewed-by: Erich Keane <erich.keane@intel.com>
Reviewed-by: Ossama Othman <ossama.othman@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>