resource-manipulation: rcs_common.so add missing depending library
rcs_common.so is also depending on liboc.so and libpthread.so, this
patch adds these dependencies to the build.
Add LIBPATH to make sure the libs will be found, this fixes a problem
in some jenkins tests.

Without this the dynamic loader will not automatically load these
libraries and it could result in unresolved dependencies at runtime.

Change-Id: I753c5ec2d35876c789d93d843542344257f44c54
Signed-off-by: jyong2.kim <jyong2.kim@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1883
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: Uze Choi <uzchoi@samsung.com>