Fixed unprocessed timed-out messages in the stack.
The portion of messages in the retransmission queue did not have the
resourceUri. As a result timed-out messages when reported back to the RI
layer did not have the resource uri which failed a check on entry into
the RI layer.

This caused the server to not de-register observers and it continued to
send observation notifications to a client that is not reachable.

Removed the check for the resource uri to fix the issue.
Added check for presence of uri before comparing it with the presence
uri so presence responses can be handled separately.

This also fixes IOT-622.

Change-Id: Ie3873d02f4163b81094ca1983e79315b85ff4751
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/1995
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Omkar Hegde <omkar.m.hegde@intel.com>
Reviewed-by: Joseph Morrow <joseph.l.morrow@intel.com>
Reviewed-by: Erich Keane <erich.keane@intel.com>