Removed C++11 specific code from the CSDK, added platform_features
The CSDK provides some nice-to-have Constructors when compiling
some POD types.  The code erronously used C++11's 'default' keyword,
which isn't available on older compilers.  Since C++11/0x is not a
requirement for the CSDK, this disables the constructors in the case
where C++11/0x is not enabled.

In order to do this, it adds a platform_features.h, which should be used
to selectively compile things based on features (rather than based on
the platform name itself)

Change-Id: Ie8b9a5e9b5d6424fd8a1f37cfdbd587f665c2638
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2036
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>