Refine codes using an enum with std atomic in DevicePresence
DevicePresence::state, a variable to keep the state of the class, is an enum type. It needs to be declared as an atomic to make it thread-safe.
std::atomic with user-defined type is not available on GCC 4.6.3, therefore std::atomic_int is used instead. The enum is scoped enum, which means it is enum class. It means it is not interchangeable with int implictly.
That's why a separate method is declared to set the value from int value. Plus, the values is controlled in the class, so it is safe to cast to int and vice versa.

Change-Id: I2689c550af293c41a9757646924f210e7fccab74
Signed-off-by: coderhyme <jhyo.kim@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2244
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Madan Lanka <lanka.madan@samsung.com>