Fix presence with ipv6 changes
Previously, the CA layer code was assuming that
all responses are always sent as unicast messages. But, in
the case of presence, it is always multicast. As a result,
the server sent presence notifications only to the ipv4
address and the clients subscribed to server's ipv6 address ignored
the notification. Adding isMulticast flag to the CAResponseInfo_t
structure and changing camessagehandler to handle multicast responses.

Change-Id: I4a0991aca6cf60b4b1ad3f0dfcda683284680adc
Signed-off-by: Omkar Hegde <omkar.m.hegde@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2252
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>