OCRepresentation: fix segmentation fault by initializing memory
dimensions was not fully initialized, but used by the code, so it could
contain arbitrary memory. If this memory was a big number this code
could allocate very much memory as this was used to create a 2 or 3
dimensional arrays.

This fixes a Segmentation fault seen in normal operation of the
garageserver and a invalid memory read and a following segmentation
fault seen with valgrind in the garageserver

Change-Id: Ifbab3d48b00bd2e3559ca3a8d79336ecf5feef6e
Signed-off-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2285
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>