Fixes the observe option size to the legal max value of 3.
The options that have variable byte length are now being
encoded to their minimal value. This patch also truncates the
option data to their maximum value allowed per the spec. One specific
example is the observer data option which receives 4 bytes of data from
the OC stack. When encoding in CoAP, the maximum value is 3 bytes.
Consequently, the CA layer which is aware of the CoAP encoding truncates
the value back to 3 bytes. Also when registering an observe, the data
length is now set to 0 using the CoAP var byte encoding which saves the
4 bytes that were just carying zero's before thsi fix.

Change-Id: I033969b407aa3426df8c85c3836206c9366fb0a2
Signed-off-by: Stephane Lejeune <stlejeun@cisco.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2351
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>