Remove a Things Manager class and expose its component classes for SDK
Previously, there was a single Thing Manager class for SDK which comprises
ThingsConfiguration, ThingsDiagnostics, GroupManager, and GroupSynchronization
classes. Now, we've decided to keep the components seperately for more
efficient maintenance for each component.

Additionally, to be aligned with OIC spec, a class name of "DiagnosticsCollection" has
been changed into "MaintenanceCollection".


Change-Id: I2b5cbf832abca3ac3ecfcab88e2153cca91a67f6
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2377
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Madan Lanka <lanka.madan@samsung.com>