Bundle isolation - use separate threads for activation
The bundle activation is now triggered in a separate thread. If one bundle
fails at activation other bundles can still succeed. When the resource
container is started and the bundles are activated, the execution flow
waits up to 10 secs to join the activators. If a bundle blocks its
activation is not completed, but other bundles are not affected.

Change-Id: Ie3f44639a4ae965941fcd1120497006830952987
Signed-off-by: Markus Jung <markus.jung@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2417
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Madan Lanka <lanka.madan@samsung.com>