Enable error responses to be sent back to the client.
The resource uri was moved from the endpoint to the uri response info
struct in IPv6.
The function sendDirectStackResponses responsible for sending out ACKs,
RESETs and error codes back to the client was not updated to take in the
resource uri.
As a result, the CA layer ate this error with a uri is NULL log and the
client did not receive any response.

Added resource Uri to the responseInfo struct to fix the problem.

Change-Id: Ic3b964c646244698c1621860922fc06b40842bc3
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2454
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jon A. Cruz <jonc@osg.samsung.com>