Removed the OCPayloadType from the payload
Previously, the CBOR container included the type of the message to make
parsing of the CBOR items easier.  This is unfortunately against the
spec, so this patch removes it and requires the caller to ensure that
the expected message type is sent.  Incorrectly formatted messages will
result in the message not passed back via the callback.

Change-Id: If1ca9eed65b5274239b670e8a2bd6cadce763c19
Signed-off-by: Erich Keane <erich.keane@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2482
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>