Add support for and handle presence of the coap ACCEPT option.
This commit adds the CBOR accept option to all requests.
When a request comes in with an accept option, the option is evaluated
against the supported encoding formats. If the format is undefined or
when CBOR is requested, CBOR will be used to encod the payload.
When the response does not carry a payload, the accept option is
ignored and the response will proceed irrespective whether the accept
format was supported or not. This allowsi for example a device that
performs a delet operation proceed first working out which format is
supported by the server.

Change-Id: I7ae3430d11f1481a91413088a959f105b216ffea
Signed-off-by: Stephane Lejeune <stlejeun@cisco.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2492
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Hauke Mehrtens <hauke.mehrtens@lantiq.com>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>