Fix broken observe requests and GETs being treated as OBSERVEs
args to some functions were validated for NULL to return error codes even though
NULLs were legal for the args.
For eg. it is legal for options to be NULL.

These incorrect validations caused early returns and observation actions
were not set to OC_OBSERVE_NO_OPTION which is 2. The field in the struct
instead defaulted to OC_OBSERVE_REGISTER = 0 due to default inits.

Fixed both by removing or replacing the incorrect validations with more
appropriate ones.

Change-Id: Ia7453227a668bb26971b2b7d84381f64151b88f9
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2659
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Sachin Agrawal <sachin.agrawal@intel.com>