Remove GroupSynchronization class in things manager
We've decided to remove the GroupSynchronization class because the current
implementation of the class is not clear to be used and does not provide
any useful functionalities to developers.
We will re-design this features more concretely and re-summit it in the future.

For Android and Tizen platform, we will summit patchsets soon.

Change-Id: Icd8116c69f9078af275aa91216c0f28db4378d32
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2661
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Madan Lanka <lanka.madan@samsung.com>