Android: Refactored SimpleClient
The app was refactored with the emphasis to readability
and cleanness. The theme has been modified to include IoTivity
logo and small android icons.
It also exercises the newly created features form the stack
such as OcResourceIdentifier.
Implemented hashCode() in OcResourceIdentifier to allow it to be
used as a key in the dictionary objects

PS 2: Removed unnecessary module reference

Change-Id: Ib1c9a56d97fa52ffe7d17c156ed228cf6e8946ef
Signed-off-by: Tim Kourt <tim.a.kourt@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2719
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>
(cherry picked from commit 7abfb5c94a9a18fbad632972f99ae4a8fb60c334)
Reviewed-on: https://gerrit.iotivity.org/gerrit/3125