/oic/res response based on the collection response structure
This is based on the new strucutre of the /oic/res as currently being defined
by Ravi. This is useful for the resource directory as it needs to include
multiple links maps inside links array.

It differs with the current /oic/res response as it includes links array with
multiple links map object.

The strucutre of the payload appears in this format:
Array
  TagsObject
    di
    baseURI
    bitmap
    ins
  LinksArray
    LinkMapObject
      href
      rt
      itf
      ins

Change-Id: Ifc61a70ede1fc44e9b2321604907d99961eea13b
Signed-off-by: Habib Virji <habib.virji@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/2701
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>