Fix crash when discovering devices & platforms.
This fixes IOT-714

Iotivity does not currently enforce the mandatory fields of device and
platform be set by applications.
As a result, servers that do NOT set device and platform info segfault
when encoding device and platform info into CBOR as CBOR expects the
fields to be present.

Modified cbor payload to work with NULL fields in device and platform
payloads.

This changed also needs to go into 1.0.0-dev.

Change-Id: Ia626422bd441069febd47a5f44dd12aa42b1d46e
Signed-off-by: Mandeep Shetty <mandeep.shetty@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/3059
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Patrick Lankswert <patrick.lankswert@intel.com>